Refactor existing adapters to use item types
